Create conditioning
The following is a variant of the code for the simple symmetric random case that runs the random walk until the first passage through 0 at a strictly positive time (as opposed to iterating through known Dyck paths of even length). This can easily be changed to deal with any increment distribution by modifying the line increase = random.randint(0, 1) * 2 - 1.